Handle None copula and expand ProbInput tests
Update the `__str__` method in ProbInput to handle
"None" copula by displaying 'Independence'.
This is a temporary solution as there is not yet
an independence copula object in UQTestFuns.

Additionally, added new unit tests for `__repr__` method
and various RNG seed behaviors to improve coverage.

This commit should resolve Issue #317.