Add cooling coffee cup model
The two-dimensional, vector-valued cooling coffee cup model
is now added as new test function to the code base.
The model simulates the temperature evolution
of a coffee cup as it cools down to an ambient temperature.
By default, the output dimensionality of the test function
is 150 (i.e., the number of time steps).
The documentation has been updated accordingly.

This commit should resolve Issue #344.